@@ -3,40 +3,39 @@ using DotVVM.Framework.Configuration;
 using DotVVM.Framework.Routing;
 using Microsoft.Extensions.DependencyInjection;
 
-namespace DotvvmApplication1
+namespace DotvvmApplication1;
+public class DotvvmStartup : IDotvvmStartup, IDotvvmServiceConfigurator
 {
-    public class DotvvmStartup : IDotvvmStartup, IDotvvmServiceConfigurator
+    // For more information about this class, visit https://dotvvm.com/docs/tutorials/basics-project-structure
+    public void Configure(DotvvmConfiguration config, string applicationPath)
     {
-        // For more information about this class, visit https://dotvvm.com/docs/tutorials/basics-project-structure
-        public void Configure(DotvvmConfiguration config, string applicationPath)
-        {
-            ConfigureRoutes(config, applicationPath);
-            ConfigureControls(config, applicationPath);
-            ConfigureResources(config, applicationPath);
-        }
+        ConfigureRoutes(config, applicationPath);
+        ConfigureControls(config, applicationPath);
+        ConfigureResources(config, applicationPath);
+    }
 
-        private void ConfigureRoutes(DotvvmConfiguration config, string applicationPath)
-        {
-            config.RouteTable.Add("Default", "", "Views/default.dothtml");
+    private void ConfigureRoutes(DotvvmConfiguration config, string applicationPath)
+    {
+        config.RouteTable.Add("Default", "", "Views/Default/default.dothtml");
+        config.RouteTable.Add("Error", "error", "Views/Error/error.dothtml");
 
-            // Uncomment the following line to auto-register all dothtml files in the Views folder
-            // config.RouteTable.AutoDiscoverRoutes(new DefaultRouteStrategy(config));    
-        }
+        // Uncomment the following line to auto-register all dothtml files in the Views folder
+        // config.RouteTable.AutoDiscoverRoutes(new DefaultRouteStrategy(config));    
+    }
 
-        private void ConfigureControls(DotvvmConfiguration config, string applicationPath)
-        {
-            // register code-only controls and markup controls
-        }
+    private void ConfigureControls(DotvvmConfiguration config, string applicationPath)
+    {
+        // register code-only controls and markup controls
+    }
 
-        private void ConfigureResources(DotvvmConfiguration config, string applicationPath)
-        {
-            // register custom resources and adjust paths to the built-in resources
-        }
+    private void ConfigureResources(DotvvmConfiguration config, string applicationPath)
+    {
+        // register custom resources and adjust paths to the built-in resources
+    }
 
-        public void ConfigureServices(IDotvvmServiceCollection options)
-        {
-            //register only services that are supported by DotVVM (otherwise, register your services in Startup.cs)
-            options.AddDefaultTempStorages("Temp");
-        }
+    public void ConfigureServices(IDotvvmServiceCollection options)
+    {
+        //register only services that are supported by DotVVM (otherwise, register your services in Startup.cs)
+        options.AddDefaultTempStorages("Temp");
     }
 }
