@@ -4,24 +4,21 @@ using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.AspNetCore;
 
-namespace DotvvmApplication1
+namespace DotvvmApplication1;
+public class Program
 {
-    public class Program
+    public static void Main(string[] args)
     {
-        public static void Main(string[] args)
-        {
-            var host = new WebHostBuilder()
-                .UseKestrel()
-                .UseContentRoot(Directory.GetCurrentDirectory())
-                .UseIISIntegration()
-                .UseStartup<Startup>()
+        var host = WebHost.CreateDefaultBuilder(args)
+            .UseIISIntegration()
+            .UseStartup<Startup>()
 #if (IncludeApplicationInsights)
-                .UseApplicationInsights()
+            .UseApplicationInsights()
 #endif
-                .Build();
+            .Build();
 
-            host.Run();
-        }
+        host.Run();
     }
 }
