@@ -8,34 +8,42 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 
-namespace DotvvmApplication1
+namespace DotvvmApplication1;
+public class Startup
 {
-    public class Startup
+    // This method gets called by the runtime. Use this method to add services to the container.
+    // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
+    public void ConfigureServices(IServiceCollection services)
     {
-        // This method gets called by the runtime. Use this method to add services to the container.
-        // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddDataProtection();
-            services.AddAuthorization();
-            services.AddWebEncoders();
+        services.AddDataProtection();
+        services.AddAuthorization();
+        services.AddWebEncoders();
 
-            services.AddDotVVM<DotvvmStartup>();
+        services.AddDotVVM<DotvvmStartup>();
+    }
+
+    // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
+    public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+    {
+        // Configure the HTTP request pipeline.
+        if (!env.IsDevelopment())
+        {
+            app.UseExceptionHandler("/Error");
+            app.UseHttpsRedirection();
+            // The default HSTS value is 30 days. You may want to change this for production scenarios, see https://aka.ms/aspnetcore-hsts.
+            app.UseHsts();
         }
 
-        // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
-        {
-            loggerFactory.AddConsole();
+        // uncomment if you want to add MVC, SignalR or other technology which uses ASP.NET Core endpoint routing
+        //app.UseRouting();
 
-            // use DotVVM
-            var dotvvmConfiguration = app.UseDotVVM<DotvvmStartup>(env.ContentRootPath);
+        // uncomment to enable authorization
+        //app.UseAuthorization();
 
-            // use static files
-            app.UseStaticFiles(new StaticFileOptions
-            {
-                FileProvider = new Microsoft.Extensions.FileProviders.PhysicalFileProvider(env.WebRootPath)
-            });
-        }
+        // use DotVVM
+        var dotvvmConfiguration = app.UseDotVVM<DotvvmStartup>(env.ContentRootPath);
+
+        // use static files
+        app.UseStaticFiles();
     }
 }
